Add hover state capability to ssb plugins
ssba_admin_panel - Generates the input rows for the hover images and
uploads them to the server
simple-share-buttons-adder - Where all the magic happens. Generates the
new database fields to store the hover image paths and renders their
images in the dom via css snippets assigned in the link generation
method.  Also, the links are given a custom style of 'ssba_share_link'
so that the theme.css file can target them.

ssba_admin.js -- some console logging in an attempt to determine what is
happening on the page when i try to upload an image.