HOVER IMAGES THAT LINK TO SHARE SITES
The user can now successfully upload base and hover images for the
social sites. When the particular service is loaded, both images come
across and display as they should.

TODO: Adjust the css styles for the buttons and get the actual share
icons to upload.  NOTE - The css styles for the buttons can be adjusted
both inside the settings page and inside the css styles in the
buttons-adder.php file. Also:  each share link has a shared style of
ssba_share_link which can be styled in the theme css file if desired.

TODO: Further cleaning needs to be done. Plugin code needs to be refactored and
cleaned up a bit. Attribution should be cleaned up as well.